backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
httpx==0.25.2
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0
gunicorn==21.2.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
